+++ src/components/icon-toolbar/icon-toolbar.test.js
@@ -1,0 +1,36 @@
+import React from 'react';
+import IconToolbar, { mapStateToProps, mapDispatchToProps } from './index';
+import { mockState, setup } from '../../utils/state.mock';
+
+describe('IconToolbar', () => {
+  it('renders without crashing', () => {
+    const wrapper = setup.mount(<IconToolbar />);
+    expect(wrapper.find('.pipeline-icon-toolbar').length).toBe(1);
+    expect(wrapper.find('.pipeline-toggle-theme').length).toBe(1);
+    expect(wrapper.find('.pipeline-toggle-labels').length).toBe(1);
+  });
+
+  it('maps state to props', () => {
+    const expectedResult = {
+      textLabels: expect.any(Boolean),
+      theme: expect.stringMatching(/light|dark/)
+    };
+    expect(mapStateToProps(mockState.lorem)).toEqual(expectedResult);
+  });
+
+  it('maps dispatch to props', () => {
+    const dispatch = jest.fn();
+
+    mapDispatchToProps(dispatch).onToggleTextLabels(false);
+    expect(dispatch.mock.calls[0][0]).toEqual({
+      textLabels: false,
+      type: 'TOGGLE_TEXT_LABELS'
+    });
+
+    mapDispatchToProps(dispatch).onToggleTheme('light');
+    expect(dispatch.mock.calls[1][0]).toEqual({
+      theme: 'light',
+      type: 'TOGGLE_THEME'
+    });
+  });
+});
